requirements.txt
# MCP Development Core: requirements.txt
# This file lists all Python dependencies for the Local Arbitrage MCP Server project.

# --- Phase 1: Core Backend & Authentication ---
# FastAPI for the core web server and API endpoints
fastapi==0.111.0
# Uvicorn as the ASGI server to run FastAPI
uvicorn[standard]==0.29.0
# Firebase Admin SDK for backend user authentication and token verification
firebase-admin==6.5.0
# Jinja2 for rendering HTML templates from the backend
Jinja2==3.1.4
# python-dotenv to manage environment variables (API keys, secrets) from a .env file
python-dotenv==1.0.1
# python-multipart for handling file uploads or form data, if needed
python-multipart==0.0.9

# --- Phase 2: Agent Development (The MCP's Brain) ---
# Google AI Python SDK to interact with the Gemini 1.5 Pro model
google-generativeai==0.7.0
# Google API Client for programmatic access to Google services like the Search API
google-api-python-client==2.134.0
google-auth-httplib2==0.2.0
# HTTPX for making robust, asynchronous HTTP requests to external APIs (e.g., scraping, third-party SEO tools)
httpx==0.27.0
# BeautifulSoup4 for parsing and scraping HTML content from competitor websites
beautifulsoup4==4.12.3
pydantic

# --- Phase 4: Deployment & Integration ---
# No additional Python libraries are required for this phase.
# The Vercel/Netlify API integration will be handled using HTTPX.
# Dockerization is an external process and does not require a library within the application itself.
# From GitHub for Sanity.io integration
-e git+https://github.com/OmniPro-Group/sanity-python.git#egg=sanity